Pawsey summer internship

Former-commit-id: 816297fba092b87058b191b06da19f2deb13f919

+++ signbreaker/create_dataset.py
@@ -1,34 +1,35 @@
 def main():
     import os
+    import sys
     import ntpath
     import shutil
-    import sys
     import argparse
     from datetime import datetime
     import random
 
-    import numpy as np
     import cv2
-    from PIL import Image, ImageChops, ImageDraw, ImageOps, ImageFilter, ImageStat, ImageEnhance, ImageFile
-    from skimage import io, color, exposure
-    import matplotlib
-    import matplotlib.pyplot as plt
+    import json
+    from collections import defaultdict
+    from PIL import ImageFile
+    from pathlib import Path
     import glob
-    import math
+    import numpy as np
 
-    from damage import no_damage, remove_quadrant, remove_hole, bullet_holes, graffiti, bend_vertical, damage_image
-    from utils import load_paths, load_files, scale_image, create_alpha, delete_background, to_png, dir_split, png_to_jpeg
+    from bg_image import BgImage
+    from damage import damage_image
+    from utils import load_paths, load_files, scale_image, delete_background, to_png, dir_split
     import manipulate
     import generate
-    from synth_image import SynthImage
-
-    #TODO: Add an example download link for a dataset of backgrounds that have no real signs on them
-
-    #TODO: Documentation in the dataset readme for what the tag of each damage type means
-
-    #TODO: Add randomisation factor somehwere so that, if desired, a new image won't be generated for every
-    #      single class+damage+transformation+(manipulation+background) combination, with the factor being
-    #      introduced somewhere within the brackets
+    
+    current_dir = os.path.dirname(os.path.realpath(__file__))
+    
+    parser = argparse.ArgumentParser(description='Create a synthetically generated traffic sign dataset.')
+    # SGTSD stands for "Synthetically Generated Traffic Sign Dataset"
+    parser.add_argument('--output_dir', type=str, default=os.path.join(current_dir, 'SGTS_Dataset'))
+    
+    args = parser.parse_args()
+    args.output_dir = os.path.abspath(args.output_dir)
+    os.chdir(current_dir)
 
     # Open and validate config file
     import yaml
@@ -47,7 +48,7 @@ def main():
         raise ValueError("Config error: must have 0 <= 'num_transform' <= 15.\n")
     if not config['man_method'] in valid_man:
         raise ValueError(f"Config error: 'man_method' must be either '{valid_man[0]}' or '{valid_man[1]}'.\n")
-    for dmg in config['damage_types']:
+    for dmg in config['num_damages']:
         if not dmg in valid_dmg:
             raise ValueError(f"Config error: '{dmg}' is an invalid damage type.\n")
 
@@ -65,6 +66,13 @@ def main():
             raise ValueError(f"Config error: must have 0.0 < 'graffiti:{g_param}' <= 1.0.\n")
     if g_params['initial'] > g_params['final']:
         raise ValueError("Config error: 'graffiti:initial' must be <= 'graffiti:final'.\n")
+    
+    if not config['reuse_data']['damage'] and config['reuse_data']['manipulate']:
+        raise ValueError("Config error: 'reuse_data:damage' must be true if 'reuse_data:manipulate' is true.\n")
+    if config['reuse_data']['damage'] and not os.path.exists(os.path.join(damaged_dir, 'damaged_data.npy')):
+        raise FileNotFoundError('Config error: damaged_data.npy does not exist. Cannot re-use data\n')
+    if config['reuse_data']['manipulate'] and not os.path.exists(os.path.join(manipulated_dir, 'manipulated_data.npy')):
+        raise FileNotFoundError('Config error: manipulated_data.npy does not exist. Cannot re-use data\n')
 
     print("Generating dataset using the 'config.yaml' configuration.\n")
 
@@ -76,7 +84,7 @@ def main():
     transformed_dir = os.path.join(base_dir, "4_Transformed")
     manipulated_dir = os.path.join(base_dir, "5_Manipulated")
     bg_dir    = "Backgrounds"
-    final_dir = "SGTS_Dataset"  # SGTSD stands for "Synthetically Generated Traffic Sign Dataset"
+    final_dir = args.output_dir
 
     # Create the output directories if they don't exist already
     if not os.path.exists(base_dir):
@@ -93,6 +101,23 @@ def main():
         shutil.rmtree(processed_dir)
     os.mkdir(processed_dir)
 
+    # Seed the random number generator
+    random.seed(config['seed'])
+    
+
+
+    ##################################
+    ###  BACKGROUND PREPROCESSING  ###
+    ##################################
+    background_paths = glob.glob(f"{bg_dir}{os.sep}**{os.sep}*.png", recursive=True) + \
+        glob.glob(f"{bg_dir}{os.sep}**{os.sep}*.jpg", recursive=True)
+    background_images = []
+    if config['detect_light_src']:
+        for ii, path in enumerate(background_paths):
+            print(f"Processing Background Images: {float(ii) / float(len(background_paths)):06.2%}", end='\r')
+            background_images.append(BgImage(path))  # Detect light sources in each image
+        print(f"Processing Background Images: 100.00%\r\n")
+
 
 
     #############################
@@ -101,15 +126,15 @@ def main():
     # Rescale images and make white backgrounds transparent
     paths = load_files(input_dir)
     for path in paths:
-        img = scale_image(path, config['sign_width']) # Rescale the image
-
-        # Remove the extension and save as a png
         _, filename = ntpath.split(path)
-        name, _ = filename.rsplit('.', 1)
+        name, extension = filename.rsplit('.', 1)
+        
+        img = scale_image(path, config['sign_width']) # Rescale the image
         save_path = os.path.join(processed_dir, name) + ".png"
         img.save(save_path)
 
-        delete_background(save_path, save_path) # Overwrite the newly rescaled image
+        if not img.mode[-1] == 'A':
+            delete_background(save_path, save_path) # Overwrite the newly rescaled image
 
     if config['final_op'] == 'process':
         return
@@ -119,33 +144,35 @@ def main():
     #########################
     ###  APPLYING DAMAGE  ###
     #########################
-    # Remove any old output and recreate the output directory
-    # reusable = config['reuse_damage']  #TODO: Implement optional damage reuse or any dir reuse (check downstream first)
-    # if not reusable:
-    #     shutil.rmtree(damaged_dir)
-    if os.path.exists(damaged_dir):
-        shutil.rmtree(damaged_dir) ##
-    #if not os.path.exists(damaged_dir):
-    os.mkdir(damaged_dir)
-    damaged_data = []
+    reusable = config['reuse_data']['damage']
+    data_file_path = os.path.join(damaged_dir, "damaged_data.npy")
+    if not reusable:
+        if os.path.exists(damaged_dir):  # Remove any old output
+            shutil.rmtree(damaged_dir)
+        os.mkdir(damaged_dir)
+        damaged_data = []
 
-    ii = 0
-    processed = load_files(processed_dir)
-    for image_path in processed:
-        print(f"Damaging signs: {float(ii) / float(len(processed)):06.2%}", end='\r')
-        damaged_data.append(damage_image(image_path, damaged_dir, config))
-        ii += 1
-    print(f"Damaging signs: 100.0%\r\n")
-    damaged_data = [cell for row in damaged_data for cell in row]  # Flatten the list
-    # else:
-    #     print("Reusing pre-existing damaged signs")
+        ii = 0
+        processed = load_files(processed_dir)
+        for image_path in processed:
+            print(f"Damaging signs: {float(ii) / float(len(processed)):06.2%}", end='\r')
+            damaged_data.append(damage_image(image_path, damaged_dir, config, background_images))
+            ii += 1
+        print(f"Damaging signs: 100.0%\r\n")
+        damaged_data = [cell for row in damaged_data for cell in row]  # Flatten the list
+        np.save(data_file_path, damaged_data, allow_pickle=True)
+    elif os.path.exists(data_file_path):
+        damaged_data = np.load(os.path.join(damaged_dir, "damaged_data.npy"), allow_pickle=True)
+        print("Reusing pre-existing damaged signs")
+    else:
+        raise FileNotFoundError(f"Error: Damaged data file does not exist - cannot reuse.\n")
 
     if config['final_op'] == 'damage':
         return
 
 
 
-    #FIXME: 'transform_type' final label keeps having value None despite synth_image.set_transformation() working
+    #FIXME(old): 'transform_type' final label keeps having value None despite synth_image.set_transformation() working
     ##################################
     ###  APPLYING TRANSFORMATIONS  ###
     ##################################
@@ -169,62 +196,42 @@ def main():
     ####################################
     ###  MANIPULATING EXPOSURE/FADE  ###
     ####################################
-    if os.path.exists(manipulated_dir):
-        shutil.rmtree(manipulated_dir)
-    os.mkdir(manipulated_dir)
+    reusable = config['reuse_data']['manipulate']
+    data_file_path = os.path.join(manipulated_dir, "manipulated_data.npy")
+    if not reusable:
+        if os.path.exists(manipulated_dir):
+            shutil.rmtree(manipulated_dir)
+        os.mkdir(manipulated_dir)
 
-    ImageFile.LOAD_TRUNCATED_IMAGES = True  #TODO: Is this line needed?
-    for bg_folders in load_paths(bg_dir):
-        to_png(bg_folders)
-
-    #TODO: This mess of a loop could probably be cleaned up?
-    for dirs in load_paths(bg_dir):
+        ImageFile.LOAD_TRUNCATED_IMAGES = True
+        for bg_folders in load_paths(bg_dir):
+            to_png(bg_folders)
+            
+        background_paths = glob.glob(f"{bg_dir}{os.sep}**{os.sep}*.png", recursive=True)
+            
         if config['man_method'] == 'exposure':
-            for background in load_paths(dirs):
-                iniitial, subd, element = dir_split(background)
-                title, extension = element.split('.')
-
-                for signp in load_paths(transformed_dir):
-                    for sign in load_paths(signp):
-                        d, s, f, e = dir_split(sign)  # Eg. s=4_Transformed, f=9, e=9_BIG_HOLE
-
-                        #if (not os.path.exists(exp_dir + subd + sep + title + sep + "SIGN_" + f + sep + e)):
-                        sign_dir = os.path.join(manipulated_dir, subd, title, "SIGN_" + f, e)
-                        if not os.path.exists(sign_dir):
-                            os.makedirs(sign_dir)
-                            #os.makedirs(exp_dir + subd + sep + title + sep + "SIGN_" + f + sep + e)
+            manipulated_data = manipulate.exposure_manipulation(transformed_data, background_paths, manipulated_dir)
+            manipulate.find_useful_signs(manipulated_data, manipulated_dir, damaged_dir)
         else:
-            for signp in load_paths(transformed_dir):
-                for sign in load_paths(signp):
-                    d,s,f,e = dir_split(sign)
+            raise NotImplementedError('Only exposure method is currently implemented')
+            # manipulated_data = manipulate.fade_manipulation(transformed_data, background_paths, manipulated_dir)
 
-                    sign_dir = os.path.join(manipulated_dir, "SIGN_" + f, e)
-                    if (not os.path.exists(sign_dir)):
-                        os.makedirs(sign_dir)
-
-    signs_paths = []
-    for p in load_paths(transformed_dir):
-        for d in load_paths(p):
-            signs_paths += load_paths(d)
-
-    background_paths = []  # Load the paths of the background images into a single list
-    for subfolder in load_paths(bg_dir):
-        background_paths += load_paths(subfolder)
-
-    #TODO: Can do checks for damage type in below functions to avoid funky results cancelling manipulation for just those types
-    
-    if config['man_method'] == 'exposure':
-        manipulated_data = manipulate.exposure_manipulation(transformed_data, background_paths, manipulated_dir)
+        # Delete SynthImage objects for any signs that were removed
+        manipulated_data[:] = [x for x in manipulated_data if os.path.exists(x.fg_path)]
+        np.save(data_file_path, manipulated_data, allow_pickle=True)
     else:
-        #TODO: Still need to do encapsulation here
-        None
-        #manipulated_data = manipulate.fade_manipulation(transformed_data, background_paths, manipulated_dir)
-
-    if config['man_method'] == 'exposure':
-        manipulate.find_useful_signs(manipulated_data, manipulated_dir, damaged_dir) #TODO: MAKE SURE TO DO THIS, DELETING SYNTH_IMAGE OBJECTS ALONG THE WAY
-
-    # Delete SynthImage objects for any signs that were removed
-    manipulated_data[:] = [x for x in manipulated_data if os.path.exists(x.fg_path)]
+        manipulated_data = np.load(data_file_path, allow_pickle=True)
+        print("Reusing pre-existing manipulated signs")
+    
+    # Prune dataset by randomly sampling from manipulated images
+    if config['prune_dataset']['prune']:
+        max_images = config['prune_dataset']['max_images']
+        images_dict = defaultdict(list)
+        for img in manipulated_data:
+            images_dict[os.path.dirname(img.fg_path)].append(img)
+        # Sample manipulated-transformed images for each background/class/damage
+        images_dict = {k:random.sample(v, max_images) for k,v in images_dict.items() if len(v) > max_images}
+        manipulated_data = [img for images_list in images_dict.values() for img in images_list]
 
     if config['final_op'] == 'manipulate':
         return
@@ -235,7 +242,18 @@ def main():
     ###  GENERATING FINAL DATA  ###
     ###############################
     images_dir = os.path.join(final_dir, "Images")
-    labels_path = os.path.join(final_dir, "labels.txt")
+    labels_format = config['annotations']['type']
+    damage_labelling = config['annotations']['damage_labelling'] == True
+    
+    # Initialise annotation files according to config parameters
+    if labels_format == 'retinanet':
+        labels_path = os.path.join(final_dir, "labels.txt")
+    elif labels_format == 'coco':
+        labels_path = os.path.join(final_dir, "_annotations.coco.json")
+        classes = [int(Path(p).stem) for p in glob.glob(f'{processed_dir}{os.path.sep}*.png')]
+        labels_dict = {'categories': [], 'images': [], 'annotations': []}
+        labels_dict["categories"] += [{"id": 0, "name": "signs", "supercategory": "none"}]
+        labels_dict["categories"] += [{'id:': c, 'name': str(c), 'supercategory': 'signs'} for c in sorted(classes)]
     about_path = os.path.join(final_dir, "generated_images_about.txt")
 
     # Clean and recreate the parent images directory
@@ -247,37 +265,36 @@ def main():
     print(f"Files to be generated: {total_gen}")
 
     ii = 0
-    with open(labels_path, "w") as labels_file:
-        for synth_image in manipulated_data:
-            print(f"Generating files: {float(ii) / float(total_gen):06.2%}", end='\r')
-            
-            c_num = synth_image.class_num
-            d_type = synth_image.damage_type
-            class_dir = os.path.join(images_dir, f"{c_num}", f"{c_num}_{d_type}")
-            # Create the directory for each class+damage combination if it doesn't already exist
-            if not os.path.exists(class_dir):
-                os.makedirs(class_dir)
-            
-            fg_path =  os.path.join(class_dir, f"{c_num}_{d_type}_{ii}")
-            temp_fg_path  = fg_path + ".png"
-            final_fg_path = fg_path + ".jpg"  # It is assumed that the final .jpg -> .png conversion step is executed
+    labels_file = open(labels_path, "w")
+    
+    for synth_image in manipulated_data:
+        print(f"Generating files: {float(ii) / float(total_gen):06.2%}", end='\r')
+        
+        c_num = synth_image.class_num
+        d_type = synth_image.damage_type
+        class_dir = os.path.join(images_dir, f"{c_num}", f"{c_num}_{d_type}")
+        # Create the directory for each class+damage combination if it doesn't already exist
+        if not os.path.exists(class_dir):
+            os.makedirs(class_dir)
+        
+        fg_path =  os.path.join(class_dir, f"{c_num}_{d_type}_{ii}")
+        final_fg_path = fg_path + ".jpg"
 
-            image = generate.new_data(synth_image, labels_file)
-            cv2.imwrite(temp_fg_path, image)
-            ii += 1
-        print(f"Generating files: 100.0%\r\n")
-
-    #TODO: Is this really necessary? Can't we save to JPEG directly?
-    ii = 0
-    class_dirs = load_paths(images_dir)
-    for class_dir in class_dirs:
-        for damage_dir in load_paths(class_dir):
-            for image in load_paths(damage_dir):
-                print(f"Converting to JPEG: {float(ii) / float(total_gen):06.2%}", end='\r')
-                if image.endswith("png"):
-                    png_to_jpeg(image)
-                ii += 1  
-    print(f"Converting to JPEG: 100.0%\r\n")
+        image = generate.new_data(synth_image)
+        if labels_format == 'retinanet':
+            synth_image.write_label_retinanet(labels_file, damage_labelling)
+        elif labels_format == 'coco':
+            synth_image.write_label_coco(labels_dict, ii, 
+                        os.path.relpath(final_fg_path, final_dir), image.shape, damage_labelling)
+        cv2.imwrite(final_fg_path, image, [int(cv2.IMWRITE_JPEG_QUALITY), 100])
+        ii += 1
+    print(f"Generating files: 100.0%\r\n")
+    
+    if labels_format == "coco":
+        labels_dict['images'] = sorted(labels_dict['images'], key=lambda x: x['id'])
+        labels_dict['annotations'] = sorted(labels_dict['annotations'], key=lambda x: x['id'])
+        json.dump(labels_dict, labels_file, indent=4)
+    labels_file.close()
 
     string = '''
     -------------------------------------
@@ -297,6 +314,10 @@ def main():
 
     with open(about_path, "w") as text_file:
         text_file.write(string)
+    
+    # Save config file with dataset
+    with open(os.path.join(final_dir, 'config.yaml'), 'w') as f:
+        yaml.dump(config, f, default_flow_style=False)
